Consolidate reminder card rendering
