+++ web-v4/src/lib/components/schedules/RemindersSection.svelte
@@ -2,8 +2,6 @@
 	import { Button } from '$lib';
 	import { umamiEvent } from '$lib/umami';
 	import { sortRemindersByDue } from '$lib/utils/reminders';
-	import { intlFormatDateUTC } from '$lib/utils/date';
-	import { formatMileage } from '$lib/utils/vehicle';
 	import { Bell } from 'lucide-svelte';
 	import ReminderCard from './ReminderCard.svelte';
 	import EmptyState from '$lib/components/EmptyState.svelte';
@@ -19,19 +17,6 @@
 
 	let completingId = $state<number | null>(null);
 	const sorted = $derived(sortRemindersByDue(reminders, vehicle.estimatedMileage));
-
-	const dueSummary = (reminder: Reminder): string => {
-		if (!reminder.reminderType) return '';
-		const parts: string[] = [];
-		if (reminder.reminderDate || reminder.date) {
-			parts.push(intlFormatDateUTC(reminder.reminderDate ?? reminder.date!));
-		}
-		if (reminder.mileage) {
-			if (parts.length) parts.push('or');
-			parts.push(formatMileage(reminder.mileage, vehicle.distanceUnit));
-		}
-		return `Due ${parts.join(' ')}`;
-	};
 </script>
 
 <div class="mb-3">
@@ -57,14 +42,15 @@
 {#if reminders.length}
 	<ul class="space-y-3">
 		{#each sorted as reminder (reminder.id)}
-			<ReminderCard
-				{reminder}
-				{vehicle}
-				dueSummary={dueSummary(reminder)}
-				completing={completingId === reminder.id}
-				oncomplete={() => (completingId = reminder.id)}
-				oncancel={() => (completingId = null)}
-			/>
+			<li>
+				<ReminderCard
+					{reminder}
+					{vehicle}
+					completing={completingId === reminder.id}
+					oncomplete={() => (completingId = reminder.id)}
+					oncancel={() => (completingId = null)}
+				/>
+			</li>
 		{/each}
 	</ul>
 {:else}
