+++ web-v4/src/routes/vehicles/[id]/reminders/+page.svelte
@@ -11,6 +11,7 @@
 	import { isReminderOverdue, sortRemindersByDue } from '$lib/utils/reminders';
 	import type { PageProps } from './$types';
 	import { pageTitle } from '$lib/utils/site';
+	import ReminderCard from '$lib/components/schedules/ReminderCard.svelte';
 
 	let { data }: PageProps = $props();
 
@@ -35,7 +36,11 @@
 		{#if reminders.length}
 			<header class="flex items-center gap-2 mb-6">
 				{#if !vehicle.retired}
-					<Button href={`/vehicles/${vehicle.id}/add?view=reminder`} class="ml-auto" {...umamiEvent('add_reminder')}>
+					<Button
+						href={`/vehicles/${vehicle.id}/add?view=reminder`}
+						class="ml-auto"
+						{...umamiEvent('add_reminder')}
+					>
 						<PlusIcon size={16} />
 						New Reminder
 					</Button>
@@ -44,69 +49,13 @@
 			<ul class="space-y-3">
 				{#each sortedReminders as reminder (reminder.id)}
 					<li>
-						<Card variant="outline" size="sm" class="gap-3">
-							<div class="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-4">
-								<div class="space-y-0">
-									<p class="font-medium text-lg flex items-center gap-3">
-										{#if isReminderOverdue(reminder, vehicle.estimatedMileage)}
-											<span class="w-2 h-2 rounded-full bg-warning shrink-0"></span>
-										{/if}
-										{reminder.notes}
-									</p>
-									<p class="text-base text-ink-500">
-										{#if !reminder.reminderType}
-											No reminder set
-										{:else}
-											Due
-											{#if reminder.reminderDate || reminder.date}
-												<span>{intlFormatDateUTC(reminder.reminderDate ?? reminder.date!)}</span>
-												{#if reminder.mileage}
-													or
-												{/if}
-											{/if}
-											{#if reminder.mileage}
-												<span>{formatMileage(reminder.mileage, vehicle.distanceUnit)}</span>
-											{/if}
-										{/if}
-									</p>
-								</div>
-								<div class="flex items-center gap-2">
-									<Button
-										size="sm"
-										href={`/vehicles/${vehicle.id}/reminders/${reminder.id}/edit`}
-										variant="ghost"
-									>
-										Edit
-									</Button>
-									{#if completingId === reminder.id}
-										<Button size="sm" variant="outline" onclick={() => (completingId = null)}
-											>Cancel</Button
-										>
-									{:else}
-										<Button
-											size="sm"
-											variant="outline"
-											onclick={() => (completingId = reminder.id)}
-											{...umamiEvent('complete_reminder')}
-										>
-											Complete
-											<ChevronRight size={16} />
-										</Button>
-									{/if}
-								</div>
-							</div>
-							{#if completingId === reminder.id}
-								<div class="border-t border-ink-200 pt-4">
-									<RecordForm
-										{vehicle}
-										classifications={data.classifications}
-										distanceUnit={vehicle.distanceUnit}
-										record={{ notes: reminder.notes, mileage: vehicle.estimatedMileage } as any}
-										action={`/vehicles/${vehicle.id}/add?/record&reminder_id=${reminder.id}`}
-									/>
-								</div>
-							{/if}
-						</Card>
+						<ReminderCard
+							{reminder}
+							{vehicle}
+							completing={completingId === reminder.id}
+							oncomplete={() => (completingId = reminder.id)}
+							oncancel={() => (completingId = null)}
+						/>
 					</li>
 				{/each}
 			</ul>
